// src/api/types.gen.ts
export type GetStandardJudgeTaskResponse = {
    model: string;
    author: string;
    input_file_url: string;
    status: 'pending' | 'instance_created' | 'running' | 'finished' | 'failed';
    created_at: number;
    instance_created_at?: number | null;
    running_at?: number | null;
    output_file_url?: string | null;
    finished_at?: number | null;
    failed_at?: number | null;
};